Rename + move KafkaProducer
